Restructered folders a bit
