+++ app/Models/SocialAccount.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace App\Models;
+
+use Illuminate\Database\Eloquent\Model;
+
+
+class SocialAccount extends Model
+{
+	protected $fillable = ['user_id', 'provider_user_id', 'provider'];
+
+	public function user()
+	{
+		return $this->belongsTo(User::class);
+	}
+}
